Add level logic that increment number of pokemons
I have to change fetching so that the function return Promise.
The reason for this is that updating of pokemonDB can be caused only by
changing pickedCards (since it's a dependency of the useEffect) and checking
if every card is picked. If that's the case, then I need to:
  1. update pokemonDB
and only after pokemonDB is updated,
  2. empty pickedCards array
otherwise condition will not be checked correctlys